# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server breachedemails server."
  version: "1.0.0"
  title: "Swagger breached emails"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "breachedemails.swagger.io"
basePath: "/v2"
tags:
- name: "breachedemail"
  description: "Everything about your breached emails"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "https"
- "http"
paths:
  /breachedemails/:
    post:
      tags:
      - "breachedemail"
      summary: "Add a breached email"
      description: "Add a breached email"
      operationId: "Add"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Add a breached email"
        required: true
        schema:
          $ref: "#/definitions/Breachedemail"
      responses:
        "200":
          description: "Created"
        "400":
          description: "Conflict"
  /breachedemails/{breachedemail}:
    get:
      tags:
      - "breachedemail"
      summary: "Confirm ok if breached email exist"
      description: "Returns OK or NotFound"
      operationId: "Get"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "breachedemail"
        in: "path"
        description: "Breachedemail to return"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
        "400":
          description: "NotFound"
    delete:
      tags:
      - "breachedemail"
      summary: "Deletes a breachedemail"
      description: "Deletes a breachedemail"
      operationId: "Delete"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "breachedemail"
        in: "path"
        description: "Breached email to delete"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
definitions:
  Breachedemail:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
    xml:
      name: "Breachedemail"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"